requirements.txt
attrs==23.1.0
Authlib==1.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
colorlog==6.7.0
cryptography==41.0.2
cssselect==1.2.0
decorator==5.1.1
distro==1.8.0
docker==6.1.3
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jsonschema==3.2.0
lxml==4.9.2
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
paramiko==3.2.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
readability-lxml==0.8.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
texttable==1.6.7
tqdm==4.65.0
types-beautifulsoup4==4.12.0.5
types-colorama==0.4.15.11
types-html5lib==1.1.11.14
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
urllib3==2.0.3
validators==0.20.0
weaviate-client==3.22.1
websocket-client==0.59.0
